pyproject.toml
[project]
name = "noi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlx-vlm>=0.4.1",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pillow-heif>=1.3.0",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "ruff>=0.15.7",
]
